MBTI_Multi-Agent_Debate_System/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
alembic==1.13.1
python-dotenv==1.0.1
langchain==0.1.16
openai==1.30.1
httpx==0.27.0
starlette==0.37.2
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
jinja2==3.1.3
tqdm==4.66.4
faiss-cpu==1.8.0
chromadb==0.4.24
sentence-transformers==2.7.0
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
loguru==0.7.2
autogen==0.9.6